modifies post comments endpoint response
including virtuals on the json also means returning other values like id (wich in json is the same as _id) and also that all the virtuals would be send on all the documents. So i decided to remove them and let the client get the url from the _id field.
Also selects the fields that are returned from the users so i don't reveal secret information unnecesarily