implements protected comment create route
this route requires a text value in the body and an authorization header that contains the JWT. The function first validates the comment text and the post id on the url, and if those are correct, authenticates the user with passport's JWT strategy.
If the user is authenticated, the comment gets saved.
(god i hate this function)