redefines Post schema
i been thinking for some time how to implement this, even to the point of having two different schemas for published and unpublished post, but i think this might be the solution.
When a post is not published, its string values don't have strict validation and default to certain values. Whereas when it's published, validations of type and length are applied to those fields, making them required.