implements comment count in Post model
a new 'comment_count' field was added to the post model. This is in order to get the comment count when fetching a post, which will reduce the ammount of requests made by the frontend app, since this data is fetched together many times.

The drawback is that maybe the comment create is a bit slower now, since the comment count on the post also has to be updated during the process.