returns refresh token as cookie on user login
The first step to implementing the refresh token in the api, is to return it to the client when it logs in. It is sent as an httpOnly cookie, so the client shouldn't be able to access it, only return it back.
The expiration time of the access token was changed to 1 minutes for testing purposes