authentication bugfix and adjustments
The app wasn't able to receive and send cookies to the client, that was fixed by configuring the cors to allow requests with credentials from the app url and removing the 'sameSite' option when setting the cookie.
Fixed a header write after send error in the refresh function.
Changed the access token expiration time to 5 minutes.