Implements user update endpoint
throught this endpoint, all the data of a user can be updated, including password and image. If the image is updated, the previous image gets deleted.
the validateId function got updated to add the user to the request if the validation went ok. This remove the need to do another user query inside the update function