updates token refresh functionality
now, each time the access token is refreshed, the user data it contains gets updated with the current user data in the database. This will be useful for cases when the user is deleted or banned, since their access will get revoked after the token expires, which is not more than 5 minutes.

Another solution would be to keep track of the refresh tokens to have a list of revoked tokens, but that also implies checking against the database on every refresh, so i decided to stick with this approach