Implements supabase image server
The images uploaded locally wouldn't get persisted if a serverless service is used to host the api, so i decided to save the images on a dedicated server. - The image uploading logic was completely refactored.
- Replaced "formidable" dependency with multer. Both accomplish the work of parsing a multipart/form-data request, but multer has a memory storage already implemented, which wasn't as straight-forward in formidable.
Basically the image files are now saved in memory until they are stored in the supabase db or rejected.